+++ src/relstorage/adapters/interfaces.py
@@ -155,6 +155,19 @@ class IDBDriver(Interface):
         like that produced by `p64`.
         """
 
+    def enter_critical_phase_until_transaction_end(self, connection, cursor):
+        """
+        Given a connection and cursor opened by this driver, cause it
+        to attempt to raise its priority and return results faster.
+
+        This mostly has meaning for gevent drivers, which may limit
+        the amount of time they spend in the hub.
+
+        This phase continues until *after* the ultimate call that
+        commits or aborts is sent.
+        """
+
+
 class IDBDriverFactory(Interface):
     """
     Information about, and a way to get, an `IDBDriver`
@@ -509,6 +522,11 @@ class IManagedDBConnection(Interface):
             on the first try.
 
         :return: The return value of ``f``.
+        """
+
+    def enter_critical_phase_until_transaction_end():
+        """
+        As for :meth:`IDBDriver.enter_critical_phase_until_transaction_end`.
         """
 
 class IManagedLoadConnection(IManagedDBConnection):